% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{as_doctags}
\alias{as_doctags}
\title{Render a converted document as DocTags}
\usage{
as_doctags(x, ...)
}
\arguments{
\item{x}{A \link{doclr_document}.}

\item{...}{Unused, for extensibility.}
}
\value{
A single string of DocTags markup.
}
\description{
Render a converted document as DocTags
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
as_doctags(doclr_convert("report.pdf", doclr_options(to_formats = "doctags")))
\dontshow{\}) # examplesIf}
}
