% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{as_html}
\alias{as_html}
\title{Render a converted document as HTML}
\usage{
as_html(x, ...)
}
\arguments{
\item{x}{A \link{doclr_document}.}

\item{...}{Unused, for extensibility.}
}
\value{
A single string of HTML.
}
\description{
Render a converted document as HTML
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
as_html(doclr_convert("report.pdf", doclr_options(to_formats = "html")))
\dontshow{\}) # examplesIf}
}
