% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doclr-package.R
\docType{package}
\name{doclr-package}
\alias{doclr}
\alias{doclr-package}
\title{doclr: R Client for the 'docling-serve' REST API}
\description{
A pure-R, tidyverse-oriented client for the 'docling-serve' document conversion REST API. Submits documents by URL or by file upload, synchronously or asynchronously, polls long-running tasks and returns results as a document object that can be rendered to Markdown, plain text, HTML, JSON or DocTags. Includes optional helpers to provision and manage a local 'docling-serve' container. No Python runtime is required.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GITHUB_USER/doclr}
  \item Report bugs at \url{https://github.com/GITHUB_USER/doclr/issues}
}

}
\author{
\strong{Maintainer}: Yves Amevoin \email{amevoin@gmail.com}

Authors:
\itemize{
  \item Yves Amevoin \email{amevoin@gmail.com}
}

}
\keyword{internal}
