% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{doclr_convert_file_async}
\alias{doclr_convert_file_async}
\title{Submit a file conversion as an asynchronous task}
\usage{
doclr_convert_file_async(
  paths,
  options = doclr_options(),
  client = doclr_client()
)
}
\arguments{
\item{paths}{Character vector of local file paths.}

\item{options}{A \code{\link[=doclr_options]{doclr_options()}} object.}

\item{client}{A \code{\link[=doclr_client]{doclr_client()}}.}
}
\value{
A \link{doclr_task} object.
}
\description{
Submit a file conversion as an asynchronous task
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
task <- doclr_convert_file_async("report.pdf")
doclr_task_status(task)
\dontshow{\}) # examplesIf}
}
