% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{doclr_convert_source_async}
\alias{doclr_convert_source_async}
\title{Submit a URL conversion as an asynchronous task}
\usage{
doclr_convert_source_async(
  urls,
  options = doclr_options(),
  client = doclr_client()
)
}
\arguments{
\item{urls}{Character vector of document URLs.}

\item{options}{A \code{\link[=doclr_options]{doclr_options()}} object.}

\item{client}{A \code{\link[=doclr_client]{doclr_client()}}.}
}
\value{
A \link{doclr_task} object.
}
\description{
Submit a URL conversion as an asynchronous task
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
task <- doclr_convert_source_async("https://arxiv.org/pdf/2408.09869")
doclr_task_status(task)
\dontshow{\}) # examplesIf}
}
