% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{is_doclr_client}
\alias{is_doclr_client}
\title{Test whether an object is a doclr client}
\usage{
is_doclr_client(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a \code{doclr_client}, otherwise \code{FALSE}.
}
\description{
Test whether an object is a doclr client
}
\examples{
is_doclr_client(doclr_client())
is_doclr_client(1)
}
