% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{is_doclr_document}
\alias{is_doclr_document}
\title{Test whether an object is a converted document}
\usage{
is_doclr_document(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a \code{doclr_document}, otherwise \code{FALSE}.
}
\description{
Test whether an object is a converted document
}
\examples{
is_doclr_document(1)
}
